+++ mq-bridge.schema.json
@@ -453,12 +453,24 @@
       ]
     },
     "DeduplicationMiddleware": {
-      "description": "Deduplication middleware configuration.\n\nPrevents duplicate messages from being processed using a Sled-backed database.\nMessages are identified by their deduplication key and removed after the TTL expires.",
+      "description": "Deduplication middleware configuration.\n\nPrevents duplicate messages from being processed using a sled, MongoDB, or SQL backend.\nMessages are identified by their deduplication key and removed after the TTL expires.",
       "type": "object",
       "properties": {
         "sled_path": {
-          "description": "Path to the Sled database directory.",
-          "type": "string"
+          "description": "Local Sled directory (legacy). Prefer `store`.",
+          "type": [
+            "string",
+            "null"
+          ],
+          "default": null
+        },
+        "store": {
+          "description": "Store URL: `sled:///path` (local), `mongodb://host/db[/collection]`, or `postgres|mysql|mariadb|sqlite://…[/table]` (shared).",
+          "type": [
+            "string",
+            "null"
+          ],
+          "default": null
         },
         "ttl_seconds": {
           "description": "Time-to-live for deduplication entries in seconds.",
@@ -469,7 +481,6 @@
       },
       "additionalProperties": false,
       "required": [
-        "sled_path",
         "ttl_seconds"
       ]
     },
